// clients/accountingService/models/AppliedTaxPolicyRecordUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AppliedTaxPolicyRecordUpdateDto = {
    tenantId?: string | null;
    enrolmentId?: string | null;
    taxPolicyId?: string | null;
    invoiceId?: string | null;
    itemId?: string | null;
    taxAmountInUSD?: number;
    taxBaseAmountInUSD?: number;
};