// clients/accountingService/models/ItemTaxPolicyRecordCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemTaxPolicyRecordCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    taxPolicyId?: string | null;
    itemPriceId?: string | null;
    itemId?: string | null;
};