// clients/accountingService/models/ItemTaxPolicyRecordUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemTaxPolicyRecordUpdateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    taxPolicyId?: string | null;
    itemPriceId?: string | null;
    itemId?: string | null;
};