// clients/accountingService/models/TaxPolicyCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaxPolicyCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    code?: string | null;
    title?: string | null;
    description?: string | null;
    isFree?: boolean;
    reduce?: boolean;
    isEnabled?: boolean;
    isDefault?: boolean;
    allowInternational?: boolean;
    hours?: number;
    days?: number;
    weeks?: number;
    months?: number;
    years?: number;
    value?: number;
    percentage?: number;
    currencyId?: string | null;
    countryId?: string | null;
    countryStateId?: string | null;
    customState?: string | null;
    customCity?: string | null;
    cityId?: string | null;
    enrolmentId?: string | null;
    tenantId?: string | null;
    zero?: boolean;
    reduced?: boolean;
    withholding?: boolean;
    fiscalAuthorityId?: string | null;
};