// clients/accountingService/models/TaxRateUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaxRateUpdateDto = {
    name?: string | null;
    rate?: number;
    value?: number;
    um?: string | null;
    unitId?: string | null;
    unitGroupId?: string | null;
    priority?: number;
    compound?: boolean;
    shipping?: boolean;
    withholding?: boolean;
    singleTransactionThreshold?: number;
    cumulativeTransactionThreshold?: number;
    fiscalAuthorityId?: string | null;
    fiscalYearId?: string | null;
    tenantId?: string | null;
    countryId?: string | null;
    taxClassId?: string | null;
    currencyId?: string | null;
    taxPolicyId?: string | null;
    enrolmentId?: string | null;
};